:recycle: Refactor
Refactored the main class
