+++ Classes/app.rb
@@ -4,22 +4,62 @@ require_relative 'book'
 require_relative 'classroom'
 require_relative 'teacher'
 require_relative 'rental'
+require_relative 'command_option'
 
 class App
   def initialize
     @books = []
     @people = []
     @rentals = []
+    @command_options = []
   end
+
+  def initialize_command_options
+    @command_options = CommandOption.new.options
+  end
+
+def list_of_options
+  puts
+  puts 'Choose an option by entering a number: '
+  puts '1 - List all books'
+  puts '2 - List all people'
+  puts '3 - Create a person'
+  puts '4 - Create a book'
+  puts '5 - Create a rental'
+  puts '6 - List all rentals for a given person id'
+  puts '7 - Exit'
+end
+
+  def option(input)
+  case input
+  when '1'
+    list_all_books
+  when '2'
+    list_all_people
+  when '3'
+    create_person
+  when '4'
+    create_book
+  when '5'
+    create_rental
+  when '6'
+    list_all_rentals
+  when '7'
+    puts "Thank You for using my School Library!"
+    exit
+  else
+    puts 'Please enter a number between 1 and 7.'
+  end
+end
 
   def console_entry_point
     puts 'Welcome to my School Library!'
     until list_of_options
       input = gets.chomp
-      if input == '7'
-        puts 'Thank You for using my School Library!'
-        break
-      end
+      # if input == ''
+      #   puts 'Thank You for using my School Library!'
+      #   break
+      # end
 
       option input
     end
